[Keymap] plattfot - Set TAPPING_TERM to zero
Had issues with the delay when holding down one of the raise/lower OSL
resulting it registering the key on the default layer instead of the
raise/lower. Given that I only use the tap-hold for changing layers I
don't need any delay for it deciding whether it's a hold or a tap.