[Keymap] plattfot - Rework layout to work with EurKey
Major change in the keymap to work with EurKey. Which relaxes some
constraints it had before when it had to take in consideration two
layouts.

With this the parenthesis can be moved to a better location instead of
being in the top right corner.

The macros needed to be moved to the function layer to make room for
the parenthesis, tilde and grave.

This also allows esc, del and rctrl to be moved to the base layer.

Only downside is that ctrl+lalt needed to be removed and instead AltGr
takes it's place. Added an extra lalt to the right side to compensate.