[Keymap] plattfot - RCtrl to thumbcluster, add back double tap
- RCtrl is now where enter used to be. Making ctrl+alt combos
symmetrical for the hands, which should be easier to hit. Found that I
was hitting enter a lot when rshift was next to nav.

- Enter is now where RShift used to be. Even though it was quite nice
to have shift on the thumb cluster, having three keys dedicated for
shift seemed a bit excessive and I still need to use the pinky
LShift/RShift when typing in Swedish as the right thumb need to press
AltGr.

- Reintroduced double tap where RCtrl was before. Don't have anything
better for the spot. And keys where you tap instead of hold down works
better on the pinkies.