[Keymap] plattfot - Move clotap to layers, add eurkey specific syms
- Move clotap to the left and right hand symbol layers. Having it
fixed on the right hand side didn't quite work. As I've split the
symbol layer in two, it felt awkward to press the clotap for symbols
on the right hand. But felt quite natural to press for the left
symbols as then the right hand pressed both the clotap and the layer.
With the move I move the clotap closer to the thumb key to make it
easier to press while holding down the layer key.

- Move the ' back to its original place as clotap is moved. Rearranged
the quote keys as I haven't really gotten used to the old location.
Some emacs keybindings got awkward to use.