[keymap] plattfot - Major layout change
Swapping the two middle thumb keys for one big one. I need to reduce
the thumb keys a bit. My thumbs get a bit tired after heavy use of the
thumb keys.

This removes the leader key and the backspace key. The backspace have
been moved to a layer above the delete key. The leader key was a bit
trickier to replace. But removed the navigation layer, swapped the
function keys with the number keys. Moved the function keys down to
the bottom keys for their respective layer. And added the navigation
to the lower layer.

The navigation with the arrow keys are now a bit more awkward to use
but should be fine as I've change a bit how I navigate in emacs, where
I use these the most.