plattfot - Move numbers to the right, function to the left
Switching over to a numpad style to make it quicker to type in
numbers. And mirroring that on the function keys to make it easier to
remember. Trying out having the 1, 2, 3, 0 on the home row, as those
keys are more frequent used. Taking inspiration from pnohty [0] with
the numpad layout.

[0] https://github.com/rayduck/pnohty