[keymap] plattfot - Add TO_BASE on both side no matter which side
Turns out that with OSL, it will be active as long as you hold it
down. Which means a TO_BASE can be placed on the layer it's switching
to without interference. Which is great as then you don't need to know
which layer you activated for it to reset.