[keymap] plattfot - Move tab and backspace to thumb cluster
Pinkies are not happy with more work, reducing the load by moving
shift up one row, then tab and backspace to the outer column of the
thumb cluster.  And moving the nav keys to where the shifts where as
it is easier to reach them than the top row.  Will take a bit of time
to get use to this.