[keymap] plattfot-mini - Kyria keymap for a 38 key setup
Removed all the keys I don't use on a kyria and ended up with this.
Moved tab and backspace to the thumb cluster as I now have two big
thumb keys.  Also swap the order of enter and backspace as it feels
more logic for left hand to push the cursor forward with the
thumb (space) and the right thumb with the same key pulling the cursor
back (backspace).

Enter is still easily accessible on the outer thumb key on the right.
And given how often I use tab, it should have its place on the thumb
cluster as well.